refactor system replication
