+++ tests/e2e/system_replication/CMakeLists.txt
@@ -2,7 +2,12 @@ find_package(gflags REQUIRED)
 
 copy_e2e_python_files(replication_experiment common.py)
 copy_e2e_python_files(replication_experiment conftest.py)
+copy_e2e_python_files(replication_experiment multitenancy_common.py)
 copy_e2e_python_files(replication_experiment multitenancy.py)
+copy_e2e_python_files(replication_experiment multitenancy_part_2.py)
+copy_e2e_python_files(replication_experiment multitenancy_part_3.py)
+copy_e2e_python_files(replication_experiment multitenancy_part_4.py)
+copy_e2e_python_files(replication_experiment multitenancy_part_5.py)
 copy_e2e_python_files(replication_experiment auth.py)
 copy_e2e_python_files_from_parent_folder(replication_experiment ".." memgraph.py)
 copy_e2e_python_files_from_parent_folder(replication_experiment ".." interactive_mg_runner.py)
